Test clean up for Sign in
